Change value of genom to short, Add utils for chromosome.

+++ Assets/Scripts/GeneticCore/Logics/ChromosomeFactory.cs
@@ -48,7 +48,7 @@ namespace GeneticCore.Logics
             {
                 if (index >= startIndex && index <= endIndex)
                 {
-                    keyPairValues.Value.Value = 1 - keyPairValues.Value.Value;
+                    keyPairValues.Value.Value = (short)(1 - keyPairValues.Value.Value);
                 }
 
                 index += 1;
